Don't use :touch on topic association on Post
This is because :touch operates whenever a record is created/updated. We only want it to touch the topic when a new post has been created. Therefore, we should use an after_create and get it to update a field called last_post_at. This field is then used in the Topic model to perform the appropriate ordering.

Related to #150, #152, #154 and #159.