Make CategoriesController inherit from Forem::ApplicationController
Fixes #186.

The issue is caused because ApplicationController is loaded first, and so Forem doesn't go up the chain looking for another ApplicationController because one's already been loaded. By being explicit like this, we make sure that the right ApplicationController (Forem::ApplicationController) is loaded, rather than the wrong one (the ApplicationController from inside the application)