Give API routes priority.
I ran into an issue where `/api/topics` was routed to
TopicsController#index with a `forum_id` of `"api"`. With this
change, `/api/` routes go to the API.

This is a temporary fix, as it creates other problems if someone
decides to start a forum called "API". We could add `api` to the
reserved words list for Friendly ID, but some installations may
already have an API forum. We could give the API routes a prefix that
can't be generated by Friendly ID's default generator (e.g., `@api`).
Or we could get rid of the prefix entirely, and just use the
`Accepts` header to determine whether a request is for the HTML UI
or the API.