Create Topic error handling.
Tried to cover all of the MUSTs and some of the SHOULDs from the JSON
API spec.

Specific errors are just returned as text messages for now.

Most of the new controller code will be extracted to a mixin module.
Likewise, the spec code can be extracted to shared behaviors.